final Commit
A website was developed as a cross platform for travel freaks. Unregistered users will be able to access the public resources. Registered users will be able to publish blogs, comment on existing published blogs, can book tour packages offered by the travel agencies on the system. Travel agencies will be able to offer tour packeges with detail information. An admin panel is there to maintain and monitor users,blogs and events. There is an option to chosing themes for users.